// planning-idsdp-pps-frontend/src/app/main/modules/project-management/models/pm-pd-selection.model.ts
export class PmPdSelectionModel {

  id: number;
  pmPdSelectionId: number;
  dppTappMasterId: number;
  dppTappMasterUuid: string;
  uuid: string;
  name: string;
  designation: string;
  nid: number;
  joiningDate: Date;
  endDate: Date;
  continuing: boolean;
  attachedFile: File;
  attachmentName: string;
  attachment: any;

}